MFD: Add JUMPtec/Kontron PLD drivers (kempld)
This patch adds the JUMPtec/Kontron PLD Linux kernel driver package

The kempld driver package provides support for the on-board PLD found on
some JUMPtec/Kontron embedded modules.
The driver is a multi function driver which consists of a core driver
and different function drivers. The function drivers are only loaded on
request if the functionality is available on the module used.

** Available drivers **

kempld-core driver:
The core driver provides the driver framework for the available function
and is necessary for all other kempld-drivers.
The driver has an internal list of supported modules and shows
information about the available PLD revision as kernel information
output.
There is additional KernelDoc API documentation available in the
Documentation/DocBook/ directory of the patched kernel source.

kempld-gpio driver:
The GPIO driver allows to access the GPIO functionality that may be
available through the PLD. Depending on the hardware none or not all
GPIOs listed in the manual may be available through this driver.
Furthermore some of the GPIOs may already be used internally and
therefore are not available through the module connector.

kempld-i2c driver:
The I2C driver allows to access the I2C functionality that may be
available through the PLD.
For the I2C bus driver there is additional documentation available in
the Documentation/i2c/busses/ directory of the patched kernel source.

kempld-wdt driver:
The watchdog driver allows to access the watchdog functionality that
may be available through the PLD.

Signed-off-by: Michael Brunner <michael.brunner@jumptec.com>
Signed-off-by: Mikhail Malyshev <mike.malyshev@gmail.com>